modules/dnsdump/requirements.txt
dnspython
aiohttp